Add an "As" method to odata.Object
When retrieving information from Balena using the odata protocol it is
possible to $extend embedded structures. The odata object is used to
store either the ID and the deferred URI to allow the caller to determine
how to obtain the data but using the As method combined with the
$extend query it is possible to retrieve all the information using a
single query.

Example:
  ```
  response, err := client.Application.GetWithQuery(ctx, "$expand=is_for__device_type&$filter=is_directly_accessible_by__user/any(dau:1%20eq%201)")
  ```

  can be used to retrieve all applications the current user has access to
  and populate the `IsForDeviceType` field on the fly.

  To retrieve this information the caller now can do

  ```
  var deviceType []*DeviceTypeResponse
  _ = response.IsForDeviceType.As(&deviceType)
  ```

	And retrieve the data already transmitted.

This can also be nested:
  ```
  response, _ := client.Application.GetWithQuery(ctx, "$expand=is_for__device_type($expand=is_of__cpu_architecture&$filter=is_directly_accessible_by__user/any(dau:1%20eq%201)+and+is_of__class+eq+'fleet'")
  ```

The raw json data is stored as an embedded object inside the odata
`Object` such that it can be unmarshalled at request.